增加TSV大文件分片录入功能,优化内存使用 修复少量bug

+++ src/main/java/com/westlake/air/swathplatform/domain/db/LibraryDO.java
@@ -30,11 +30,11 @@ public class LibraryDO extends BaseDO {
 
     String instrument;
 
-    Integer proteinCount;
+    Long proteinCount;
 
-    Integer peptideCount;
+    Long peptideCount;
 
-    Integer transitionCount;
+    Long transitionCount;
 
     Date createDate;
